Refactor Applications exporter into a class
- Change to use in-memory files instead of exports writing to disk.
- Move export functionality from ApplicationRound list view to change view (Always only one ApplicationRound was exported)
- Now the exporter has a clearer structure and exporters added in the future can follow already established patterns
